Allow 'build_meta' to be importable
